routes in utils and check platform to know for mobile or desktop
