+++ client/lib/common/pages/user_pages/signup_page.dart
@@ -6,6 +6,9 @@ import 'package:qfqq/common/providers/router_provider.dart';
 import 'package:qfqq/common/services/auth_service.dart';
 import 'package:go_router/go_router.dart';
 import 'package:qfqq/common/theme/styles.dart';
+import 'package:qfqq/common/widgets/reusables/adaptive_layout.dart';
+import 'package:qfqq/common/pages/user_pages/layouts/auth_desktop_layout.dart';
+import 'package:qfqq/common/pages/user_pages/layouts/auth_mobile_layout.dart';
 import 'package:qfqq/generated/l10n.dart';
 
 class SignupPage extends ConsumerStatefulWidget {
@@ -52,31 +55,19 @@ class _SignupPageState extends ConsumerState<SignupPage> {
 
   @override
   Widget build(BuildContext context) {
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        if (constraints.maxWidth > 600) {
-          return _DesktopLayout(
-            formKey: _signupFormKey,
-            onSubmit: _submitSignup,
-            error: error,
-            onEmailSaved: (v) => newUser.email = v ?? '',
-            onFirstNameSaved: (v) => newUser.firstName = v ?? '',
-            onLastNameSaved: (v) => newUser.lastName = v ?? '',
-            onPasswordSaved: (v) => newPassword = v ?? '',
-            onConfirmPasswordSaved: (v) => confirmPassword = v ?? '',
-          );
-        }
-        return _MobileLayout(
-          formKey: _signupFormKey,
-          onSubmit: _submitSignup,
-          error: error,
-          onEmailSaved: (v) => newUser.email = v ?? '',
-          onFirstNameSaved: (v) => newUser.firstName = v ?? '',
-          onLastNameSaved: (v) => newUser.lastName = v ?? '',
-          onPasswordSaved: (v) => newPassword = v ?? '',
-          onConfirmPasswordSaved: (v) => confirmPassword = v ?? '',
-        );
-      },
+    final form = _SignupForm(
+      formKey: _signupFormKey,
+      onSubmit: _submitSignup,
+      error: error,
+      onEmailSaved: (v) => newUser.email = v ?? '',
+      onFirstNameSaved: (v) => newUser.firstName = v ?? '',
+      onLastNameSaved: (v) => newUser.lastName = v ?? '',
+      onPasswordSaved: (v) => newPassword = v ?? '',
+      onConfirmPasswordSaved: (v) => confirmPassword = v ?? '',
+    );
+    return AdaptiveLayout(
+      mobile: (_) => AuthMobileLayout(child: form),
+      desktop: (_) => AuthDesktopLayout(child: form),
     );
   }
 }
@@ -150,110 +141,6 @@ class _SignupForm extends StatelessWidget {
           ),
         ],
       ),
-    );
-  }
-}
-
-class _MobileLayout extends StatelessWidget {
-  final GlobalKey<FormState> formKey;
-  final VoidCallback onSubmit;
-  final AccountError error;
-  final ValueSetter<String?> onEmailSaved;
-  final ValueSetter<String?> onFirstNameSaved;
-  final ValueSetter<String?> onLastNameSaved;
-  final ValueSetter<String?> onPasswordSaved;
-  final ValueSetter<String?> onConfirmPasswordSaved;
-
-  const _MobileLayout({
-    required this.formKey,
-    required this.onSubmit,
-    required this.error,
-    required this.onEmailSaved,
-    required this.onFirstNameSaved,
-    required this.onLastNameSaved,
-    required this.onPasswordSaved,
-    required this.onConfirmPasswordSaved,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(24),
-      child: _SignupForm(
-        formKey: formKey,
-        onSubmit: onSubmit,
-        error: error,
-        onEmailSaved: onEmailSaved,
-        onFirstNameSaved: onFirstNameSaved,
-        onLastNameSaved: onLastNameSaved,
-        onPasswordSaved: onPasswordSaved,
-        onConfirmPasswordSaved: onConfirmPasswordSaved,
-      ),
-    );
-  }
-}
-
-class _DesktopLayout extends StatelessWidget {
-  final GlobalKey<FormState> formKey;
-  final VoidCallback onSubmit;
-  final AccountError error;
-  final ValueSetter<String?> onEmailSaved;
-  final ValueSetter<String?> onFirstNameSaved;
-  final ValueSetter<String?> onLastNameSaved;
-  final ValueSetter<String?> onPasswordSaved;
-  final ValueSetter<String?> onConfirmPasswordSaved;
-
-  const _DesktopLayout({
-    required this.formKey,
-    required this.onSubmit,
-    required this.error,
-    required this.onEmailSaved,
-    required this.onFirstNameSaved,
-    required this.onLastNameSaved,
-    required this.onPasswordSaved,
-    required this.onConfirmPasswordSaved,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Expanded(
-          child: Container(
-            color: Theme.of(context).colorScheme.primary,
-            child: Center(
-              child: Text(
-                'QFQQ',
-                style: Theme.of(context).textTheme.displayLarge?.copyWith(
-                  color: Theme.of(context).colorScheme.onPrimary,
-                ),
-              ),
-            ),
-          ),
-        ),
-        Expanded(
-          child: Center(
-            child: ConstrainedBox(
-              constraints: const BoxConstraints(maxWidth: 400),
-              child: Padding(
-                padding: const EdgeInsets.all(32),
-                child: SingleChildScrollView(
-                  child: _SignupForm(
-                    formKey: formKey,
-                    onSubmit: onSubmit,
-                    error: error,
-                    onEmailSaved: onEmailSaved,
-                    onFirstNameSaved: onFirstNameSaved,
-                    onLastNameSaved: onLastNameSaved,
-                    onPasswordSaved: onPasswordSaved,
-                    onConfirmPasswordSaved: onConfirmPasswordSaved,
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ),
-      ],
     );
   }
 }
