+++ client/lib/mobile/mobile_router.dart
@@ -18,6 +18,7 @@ import 'package:qfqq/common/pages/user_pages/signup_page.dart';
 import 'package:qfqq/common/pages/agenda_list_page.dart';
 import 'package:qfqq/common/widgets/scaffolds/auth_page_scaffold.dart';
 import 'package:qfqq/common/widgets/scaffolds/mobile_page_scaffold.dart';
+import 'package:qfqq/common/utils/route_titles.dart';
 import 'package:qfqq/generated/l10n.dart';
 
 final GoRouter mobileRouter = GoRouter(
@@ -116,41 +117,18 @@ final GoRouter mobileRouter = GoRouter(
 
 String? _getTitleForRoute(BuildContext context, GoRouterState state) {
   final loc = S.of(context);
-  final fullPath = state.fullPath;
 
-  if (fullPath == null) return null;
-
-  switch (fullPath) {
-    case '/':
-      return loc.homePageTitle;
+  switch (state.fullPath) {
     case '/agenda':
       final agenda = state.extra as MeetingAgenda?;
       return agenda == null
           ? '${loc.agendaPageTitleAppBar} - ${loc.commonCreate}'
           : '${loc.agendaPageTitleAppBar} - ${loc.commonUpdate}';
-    case '/agendas':
-      return loc.agendasListPageTitle;
-    case '/agendas/:id':
-      return loc.agendaPageTitleAppBar;
-    case '/decisions':
-      return loc.decisionsListPageTitle;
-    case '/decisions/:id':
-      return loc.decisionViewPageTitle;
-    case '/projects':
-      return loc.projectPageTitle;
     case '/project/creation':
       return loc.projectModificationPageTitle;
     case '/project/:id':
       return loc.projectViewPageTitle;
-    case '/profile':
-      return loc.profilePageTitle;
-    case '/login':
-      return loc.loginPageTitle;
-    case '/signup':
-      return loc.signupPageTitle;
-    case '/forgotten-password':
-      return loc.forgottenPasswordPageTitle;
     default:
-      return null;
+      return getTitleForRoute(context, state);
   }
 }
